+++ Examples/Example25.ps1
@@ -12,9 +12,8 @@ $example25 = Document -Name 'PScribo Example 25' {
         definition of a custom table style. The custom "BlueZebra" table style is then
         applied to a -List view table of the 11th to last service.
 
-        NOTE: List view tables do not apply the -AlternateRowStyle styling. Therefore,
-        the -AlternateRowStyle property only needs to be defined if the style is
-        applied to both "standard" and -List view tables.
+        NOTE: List view tables do apply the -AlternateRowStyle styling, but this style
+        defaults to the -RowStyle parameter (unless explicitly defined).
     #>
     Style -Name 'BlueZebraHeading' -Bold -Color 039 -Font 'Segoe UI' -BackgroundColor E8EDFF
     Style -Name 'BlueZebraRow' -Color 669 -Font 'Lucida Sans Unicode'